python-wallet/requirements.txt
flask
requests
cryptography
eth-account
qrcode[pil]
flask-wtf
base58
psycopg2-binary
PyJWT
mnemonic